move unbundleAs to utils
