+++ src/components/claims/index.tsx
@@ -4,7 +4,7 @@ import { listRequest } from "../../api/api";
 import { navigate } from "raviger";
 import { resoureType } from "../../utils/StringUtils";
 import Loading from "../common/Loading";
-import { unbundleAs } from "../eligibility";
+import unbundleAs from "../../utils/unbundleAs";
 
 export interface IAdditionalInfo {
   status: "Pending" | "Approved" | "Rejected";
@@ -61,7 +61,7 @@ export type ClaimDetail = {
     patient: object;
     coverage: object;
     claim: object;
-  }
+  };
 };
 
 export function currencyObjToString({
@@ -98,7 +98,7 @@ export function claimsMapper(claim: any): ClaimDetail {
     coverage: unbundleAs(claim.payload, "Coverage").resource,
     claim: unbundleAs(claim.payload, "Claim").resource,
   };
- 
+
   const name = resources.patient.name[0].text;
   const gender = resources.patient.gender;
 
@@ -121,7 +121,7 @@ export function claimsMapper(claim: any): ClaimDetail {
     ...parseAdditionalInfo(claim.additional_info),
     expiry: "2023-12-12",
     status: claim.status,
-    resources
+    resources,
   };
 }
